Made indexing function taking and returning longs

+++ jgrapht-unimi-dsi/src/main/java/org/jgrapht/sux4j/SuccinctDirectedGraph.java
@@ -62,6 +62,13 @@ import it.unimi.dsi.sux4j.util.EliasFanoMonotoneLongBigList;
  * {@link Integer} as edge type. Please read the {@linkplain org.jgrapht.sux4j class documentation}
  * for more information.
  *
+ * <p>
+ * For convenience, and as a compromise with the approach of {@link SuccinctIntDirectedGraph}, this
+ * class provides methods {@link org.jgrapht.sux4j.SuccinctDirectedGraph#getEdgeFromIndex(int)
+ * getEdgeFromIndex()} and
+ * {@link org.jgrapht.sux4j.SuccinctDirectedGraph#getIndexFromEdge(it.unimi.dsi.fastutil.ints.IntIntPair)
+ * getIndexFromEdge()} that map bijectively the edge set into a contiguous set of longs.
+ *
  * @author Sebastiano Vigna
  * @see SuccinctIntDirectedGraph
  */
@@ -229,13 +236,13 @@ public class SuccinctDirectedGraph
      * @return the index associated with the edge, or &minus;1 if the edge is not part of the graph.
      * @see #getEdgeFromIndex(int)
      */
-    public int getIndexFromEdge(final IntIntPair e)
+    public long getIndexFromEdge(final IntIntPair e)
     {
         final int source = e.firstInt();
         final int target = e.secondInt();
         if (source < 0 || source >= n || target < 0 || target >= n)
             throw new IllegalArgumentException();
-        return (int) successors.indexOfUnsafe(((long) source << sourceShift) + target);
+        return successors.indexOfUnsafe(((long) source << sourceShift) + target);
     }
 
     /**
@@ -245,7 +252,7 @@ public class SuccinctDirectedGraph
      * @return the pair with index {@code i}.
      * @see #getIndexFromEdge(IntIntPair)
      */
-    public IntIntPair getEdgeFromIndex(final int i)
+    public IntIntPair getEdgeFromIndex(final long i)
     {
         if (i < 0 || i >= m)
             throw new IllegalArgumentException();
